requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
flask-cors==4.0.0
eventlet==0.33.3
gunicorn==21.2.0
psycopg2-binary==2.9.10
requests==2.31.0
pandas==2.3.0
xlsxwriter==3.2.5
reportlab==4.4.2
python-dotenv
Pillow

